azure_transport_pipeline/requirements.txt
requests
pandas
azure-storage-blob
python-dotenv